backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Bcrypt==1.0.1
Flask-CORS==4.0.0
xrpl-py==2.6.0 